+++ src/roiextractors/extractors/suite2p/__init__.py
@@ -1,1 +1,13 @@
+"""A segmentation extractor for Suite2p.
+
+Modules
+-------
+suite2psegmentationextractor
+    A segmentation extractor for Suite2p.
+
+Classes
+-------
+Suite2pSegmentationExtractor
+    A segmentation extractor for Suite2p.
+"""
 from .suite2psegmentationextractor import Suite2pSegmentationExtractor